data/requirements.txt
numpy>=1.24.0
pillow>=10.0.0
pyephem>=4.1.5
astropy>=5.3.0
psutil>=5.9.0
tqdm>=4.65.0
opencv-python>=4.8.0
matplotlib>=3.7.0
scipy>=1.11.0